requirements.txt
torch>=2.6.0
numpy>=2.1.3
tqdm>=4.67.1
more-itertools>=10.6.0
transformers>=4.50.2
openai-whisper @ git+https://github.com/openai/whisper.git
numba>=0.61.0
tiktoken>=0.9.0
ffmpeg-python
